--- app/client/src/IDE/Interfaces/EditorState.ts
@@ -1,6 +1,0 @@
-export enum EditorState {
-  DATA = "DATA",
-  EDITOR = "EDITOR",
-  SETTINGS = "SETTINGS",
-  LIBRARIES = "LIBRARIES",
-}
