+++ app/client/src/ce/pages/Editor/IDE/EditorPane/JS/utils/getJSUrl.test.ts
@@ -2,7 +2,7 @@ import { getJSUrl } from "./getJSUrl";
 import urlBuilder from "ee/entities/URLRedirect/URLAssembly";
 import type { FocusEntityInfo } from "navigation/FocusEntity";
 import { FocusEntity } from "navigation/FocusEntity";
-import { EditorState } from "IDE/Interfaces/EditorState";
+import { EditorState } from "IDE/enums";
 
 describe("getJSUrl", () => {
   urlBuilder.setCurrentBasePageId("0123456789abcdef00000000");
