+++ app/client/src/pages/Editor/CanvasLayoutConversion/hooks/useShowSnapShotBanner.ts
@@ -1,5 +1,5 @@
 import { FEATURE_FLAG } from "ee/entities/FeatureFlag";
-import { EditorState } from "IDE/Interfaces/EditorState";
+import { EditorState } from "IDE/enums";
 import { getReadableSnapShotDetails } from "layoutSystems/autolayout/utils/AutoLayoutUtils";
 import { useCurrentAppState } from "pages/Editor/IDE/hooks/useCurrentAppState";
 import { useSelector } from "react-redux";
