+++ app/client/src/pages/Editor/IDE/Layout/hooks/useGridLayoutTemplate.ts
@@ -7,7 +7,7 @@ import { useCurrentAppState } from "../../hooks/useCurrentAppState";
 import { getPropertyPaneWidth } from "selectors/propertyPaneSelectors";
 import { previewModeSelector } from "selectors/editorSelectors";
 import { getIDEViewMode } from "selectors/ideSelectors";
-import { EditorState } from "IDE/Interfaces/EditorState";
+import { EditorState } from "IDE/enums";
 import { EditorEntityTab, EditorViewMode } from "IDE/Interfaces/EditorTypes";
 import {
   APP_SETTINGS_PANE_WIDTH,
