+++ app/client/src/pages/Editor/IDE/hooks/useCurrentAppState.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { EditorState } from "IDE/Interfaces/EditorState";
+import { EditorState } from "IDE/enums";
 import { useLocation } from "react-router";
 import { identifyEntityFromPath } from "navigation/FocusEntity";
