+++ app/client/src/pages/Editor/WidgetsEditor/components/CodeModeTooltip.tsx
@@ -6,7 +6,7 @@ import { getWidgetSelectionBlock } from "selectors/ui";
 import { retrieveCodeWidgetNavigationUsed } from "utils/storage";
 import { CANVAS_VIEW_MODE_TOOLTIP, createMessage } from "ee/constants/messages";
 import { useCurrentAppState } from "pages/Editor/IDE/hooks/useCurrentAppState";
-import { EditorState } from "IDE/Interfaces/EditorState";
+import { EditorState } from "IDE/enums";
 
 /**
  * CodeModeTooltip
