+++ app/client/src/pages/Editor/WidgetsEditor/components/WidgetEditorNavigation.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef, useEffect, useRef, useState } from "react";
 import NavigationPreview from "./NavigationPreview";
-import { EditorState } from "IDE/Interfaces/EditorState";
+import { EditorState } from "IDE/enums";
 import { useCurrentAppState } from "pages/Editor/IDE/hooks/useCurrentAppState";
 import {
   getAppSettingsPaneContext,
